+++ src/ckodo/encoder.hpp
@@ -9,6 +9,8 @@
 
 namespace kodo
 {
+
+    /// Base class for an encoder implementation
     struct encoder : public coder
     {
         virtual uint32_t encode(uint8_t *) = 0;
@@ -20,6 +22,11 @@ namespace kodo
             uint32_t index, const uint8_t* data, uint32_t size) = 0;
 
         virtual ~encoder(){}
+
+        virtual bool is_systematic() const = 0;
+        virtual bool is_systematic_on() const = 0;
+        virtual void set_systematic_on() = 0;
+        virtual void set_systematic_off() = 0;
     };
 }